move library code to separate package
this prevents a warning/error due to slicer thinking that
isosurfaces.py was trying to be a scripted module.

thanks @jcfr
